Finish up some stuff

+++ Utilities/Resources/MineralTypes/MiningWeight.cs
@@ -2,5 +2,5 @@ using Godot;
 [GlobalClass]
 public partial class MiningWeight : Resource {
 	[Export] MineralProfile mineralProfile;
-    [Export] public float weight;
+    [Export] public double weight;
 }
